Implemented a simple fire animation screen effect using a new particle emitter system.
Reworked textures to now contain a TextureComponent object for the texture, which is meant to allow for ease of using the new TextureSet object for animations using a single texture atlas. This means that the old method of specifying arrays of textures in texture_directory.json and having them be processed and split up at load time is now technically deprecated.