MASSIVE REWORK
Merged many changes from DDPC 2's development in order to facilitate a proper model. The engine should now be used as a compiled JAR which is then added to game projects as a library.

Also fixed loading resource files. Might need some more work to make the system robust when exporting a game project as a runnable executable.