Reworked mouse input for the GUI.
Clicks are now processed as events passed down the layer stack until they are consumed by a component. GUIComponents can override the consumeClick() method to define custom logic for whether they consume a given MouseClickEvent or not.

Also fixed a bug with the Stack implementation that caused elements to be reversed when polled using getElements()