Massive GUIText overhaul, added a TextInputComponent
The amount of changes to the GUIText system in this commit is huge, but off the top of my head (and from looking over the diff):

- Fixed how text meshes are generated to handle multiple spaces between words
- Word objects now store the number of spaces preceding them
- The DisplayManager can now perform the inverse pixel to screen operation to get pixel coordinate values from screen values
- Fixed the GUI focus system, when focus changes the previously focused component will now lose focus
- GUIComponents remember if they're focused or not, this can be queried using isFocused() when needed (such as in update())

The TextInputComponent is not finished and input is still buggy with long strings and multiple spaces. It also seems to offset itself to the left when in SCROLL mode for every unique run of spaces, and I haven't figured out why. The positioning of the cursor is only correct for vertically centered, single-line text inputs.

There's more work to be done, but for now the TextInputComponent is sufficient for inputting short strings in a single-line manner, as long as not too many spaces are used.